# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat API",
    "description": "Multi-LLM chat service API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/bot_answer_history/bot/{bot_uuid}": {
      "get": {
        "operationId": "listBotAnswerHistory",
        "summary": "List answer history for a bot",
        "tags": [
          "Bot history"
        ],
        "parameters": [
          {
            "name": "bot_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/botAnswerHistoryPageResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/bot_answer_history/bot/{bot_uuid}/count": {
      "get": {
        "operationId": "countBotAnswerHistory",
        "summary": "Count answer history entries for a bot",
        "tags": [
          "Bot history"
        ],
        "parameters": [
          {
            "name": "bot_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/countHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/chat_file/{uuid}/list": {
      "get": {
        "operationId": "listChatFiles",
        "summary": "List files for a chat session",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    },
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "id",
                      "name"
                    ]
                  },
                  "type": "array"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/download/{id}": {
      "delete": {
        "operationId": "deleteChatFile",
        "summary": "Delete a chat file",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "downloadChatFile",
        "summary": "Download a chat file",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/upload": {
      "post": {
        "operationId": "uploadChatFile",
        "summary": "Upload a chat file",
        "tags": [
          "Files"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/uploadFileContractRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/fileUploadHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "botAnswerHistoryPageResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "items": {
            "items": {
              "properties": {
                "id": {
                  "type": "integer"
                },
                "botUuid": {
                  "type": "string"
                },
                "userId": {
                  "type": "integer"
                },
                "prompt": {
                  "type": "string"
                },
                "answer": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "tokensUsed": {
                  "type": "integer"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "userUsername": {
                  "type": "string"
                },
                "userEmail": {
                  "type": "string"
                }
              },
              "additionalProperties": false,
              "type": "object",
              "required": [
                "id",
                "botUuid",
                "userId",
                "prompt",
                "answer",
                "model",
                "tokensUsed",
                "createdAt",
                "updatedAt",
                "userUsername",
                "userEmail"
              ]
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "countHTTPResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "count": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "count"
        ]
      },
...
      "fileUploadHTTPResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "url": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "size": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url",
          "name",
          "type",
          "size"
        ]
      },
...
      "uploadFileContractRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "contentEncoding": "base64"
          },
          "session-uuid": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "file",
          "session-uuid"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Browser JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}